libWsprPico target, cppcheck, strict error checking, and fixes from that

+++ src/WsprUtl.h
@@ -171,7 +171,7 @@ public:
         {
             size_t len = Len();
 
-            for (size_t i = len - 1; i >= 0; --i)
+            for (int i = len - 1; i >= 0; --i)
             {
                 if (buf_[i] == ' ')
                 {
